+++ src/settingsdialog.h
@@ -73,6 +73,7 @@ private:
     QCheckBox *m_sessionLogging;
     QCheckBox *m_autoCopy;
     QCheckBox *m_confirmMultilinePaste;
+    QCheckBox *m_confirmCloseWithProcesses;
     QLineEdit *m_editorCmd;
     QLineEdit *m_imagePasteDir;
     QComboBox *m_tabTitleFormat;